Upgrade to version 0.4.0
