--- tree_rs/src/item/default.rs
@@ -1,85 +1,0 @@
-use crate::item::DisplayBrightGreen;
-use crate::WalkDirs;
-use crate::stat::total::Totals;
-use std::ffi::OsString;
-use std::fs::{self, DirEntry, FileType};
-use std::io;
-use std::io::Write;
-use std::path::PathBuf;
-
-use crate::flag::Flags;
-use crate::handle::OutputHandler;
-use crate::loc::PrintLocation;
-use crate::stat::{Stat, total};
-use crate::tree::Tree;
-
-use super::DisplayOsString;
-
-// pub trait Visitor {
-//     fn visit_item() {}
-// }
-
-pub struct ItemCollector {
-    pub name: OsString,
-    pub path: PathBuf,
-    pub depth: i32,
-    pub file_type: FileType,
-    pub size: u64,
-}
-
-impl ItemCollector {
-    pub fn new(entry: &DirEntry, depth: &i32) -> io::Result<ItemCollector> {
-        let full_path = entry.path();
-        let metadata = fs::symlink_metadata(&full_path)?;
-        let file_type = entry.file_type()?;
-
-        Ok(ItemCollector {
-            name: full_path
-                .file_name()
-                .map(|os_str| os_str.to_os_string())
-                .unwrap_or_else(|| "Invalid full-path".into()),
-            path: full_path.clone(),
-            depth: *depth,
-            file_type,
-            size: metadata.len(),
-        })
-    }
-
-    // fn visit_item() {}
-
-    // pub fn get_item() {
-
-    // }
-
-    pub fn get_item<'a>(
-        &self,
-        flags: &Flags,
-        handler: &mut OutputHandler,
-        total: &'a mut Totals,
-        mut tree: Tree, // Remove the &mut here
-    ) {
-        if self.file_type.is_dir() {
-            // Avoid ANSI color if printing in a file,
-            // but include ANSI when printing to the terminal.
-            if flags.loc == PrintLocation::File {
-                writeln!(handler, "{}", DisplayOsString(&self.name)).unwrap_or_default();
-            } else {
-                writeln!(handler, "{}", DisplayBrightGreen(&self.name)).unwrap_or_default();
-            }
-    
-            total.directories += 1;
-    
-            tree.reach += 1;
-    
-            let mut walker = WalkDirs::new(&mut tree, &self.path, total, handler, flags);
-    
-            walker.walk_dirs();
-        } else {
-            writeln!(handler, "{}", DisplayOsString(&self.name)).unwrap_or_default();
-            total.files += 1;
-        }
-    
-        total.size += self.size;
-    }
-    
-}
