--- tree_rs/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    tree_rs::arg_builder();
-}
